+++ contrib/opencensus-ext-httplib/CHANGELOG.md
@@ -16,7 +16,7 @@ Released 2019-08-26
 ## 0.7.1
 Released 2019-08-06
 
-- Support exporter changes in `opencensus>=0.7.0`
+  - Support exporter changes in `opencensus>=0.7.0`
 
 ## 0.1.3
 Released 2019-05-31